Split out Gutter/Error
